+++ ChangeLog.md
@@ -2,6 +2,8 @@
 
 ## 0.1.3.0
 
+- Rewrite the aggregation functionality and use QuickCheck to make sure that
+  this works as expected.
 - Change the constructor for `AggregationTimes` to include the type of event
   that is being aggregated at that point in time.